added burning stamina for running

+++ Assets/Scripts/UIElements/StaminaBar.cs
@@ -5,13 +5,13 @@ public class StaminaBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxStamina(int stamina)
+    public void SetMaxStamina(float stamina)
     {
         slider.maxValue = stamina;
         slider.value = stamina;
     }
 
-    public void SetStamina(int stamina)
+    public void SetStamina(float stamina)
     {
         slider.value = stamina;
     }
